+++ .changeset/slow-turkeys-sell.md
@@ -1,0 +1,9 @@
+---
+'@obosbbl/grunnmuren-react': minor
+---
+
+Several changes to `<Select>` to make it actually usable:
+
+ * fix padding and focus styles
+ * add support `size` prop
+ * add support for `label`, `description` and `error` props
