+++ packages/react/src/Select/Select.tsx
@@ -1,32 +1,91 @@
-import { forwardRef } from 'react';
-import { cx } from '@/utils';
-import { ChevronDown } from '@obosbbl/grunnmuren-icons';
+import { forwardRef, useRef } from 'react';
+import { cx, composeRefs } from '@/utils';
+import { useFallbackId } from '@/hooks';
+import { FormLabel, FormHelperText, FormErrorMessage } from '..';
+import { useFormControlValidity } from '../hooks';
+import { SelectPlain } from './SelectPlain';
 
-export interface SelectProps extends React.ComponentPropsWithoutRef<'select'> {
+export interface SelectProps
+  extends Omit<React.ComponentPropsWithoutRef<'select'>, 'size'> {
+  /**
+   * Collection of <option />-elements
+   */
   children: React.ReactNode;
-  /** Render select as invalid. Sets `aria-invalid` to true */
-  isInvalid?: boolean;
+
+  /** Help text for the form control */
+  description?: React.ReactNode;
+
+  /** Disables the built in HTML5 validation. If using custom validation for an entire form, consider setting `noValidate` on the form element instead. @default false */
+  disableValidation?: boolean;
+
+  /** Error message for the form control */
+  error?: string;
+
+  /**  Label for the form control */
+  label: string;
+
+  /**
+   * Changes font-size, padding and gaps
+   * @default medium
+   */
+  size?: 'medium' | 'small';
 }
 
 export const Select = forwardRef<HTMLSelectElement, SelectProps>(
   (props, ref) => {
-    const { children, className, isInvalid, ...rest } = props;
+    const {
+      description,
+      error,
+      id: idProp,
+      label,
+      disableValidation = false,
+      ...rest
+    } = props;
+
+    const ownRef = useRef(null);
+
+    const { validity, validationMessage } = useFormControlValidity(
+      ownRef,
+      !disableValidation,
+    );
+
+    const id = useFallbackId(idProp);
+    const helpTextId = id + 'help';
+    const errorMsgId = id + 'err';
+
+    const errorMsg = error || validationMessage;
 
     return (
-      <div className={cx('relative', className)}>
-        <select
-          {...rest}
-          className={cx(
-            'w-full cursor-pointer appearance-none rounded-lg border border-b-[3px] border-solid bg-white px-4 py-3.5 focus:-mt-0.5 focus:-ml-0.5 focus:border-[3px] focus:shadow focus:outline-none',
-            isInvalid
-              ? 'border-red focus:border-red'
-              : 'focus:border-blue-dark border-black',
-          )}
-          ref={ref}
+      <div className="grid gap-2">
+        <FormLabel
+          htmlFor={id}
+          isRequired={props.required}
+          isInvalid={!!error || validity === 'invalid'}
         >
-          {children}
-        </select>
-        <ChevronDown className="pointer-events-none absolute top-4 right-4" />
+          {label}
+        </FormLabel>
+
+        {description && (
+          <FormHelperText id={helpTextId}>{description}</FormHelperText>
+        )}
+
+        <SelectPlain
+          id={id}
+          ref={composeRefs(ownRef, ref)}
+          {...rest}
+          // for accessibility reasons these cannot be overriden
+          isInvalid={!!error || validity === 'invalid'}
+          aria-describedby={
+            cx({
+              [errorMsgId]: errorMsg,
+              [helpTextId]: description,
+            }) || undefined
+          }
+        />
+
+        {errorMsg && (
+          <FormErrorMessage id={errorMsgId}>{errorMsg}</FormErrorMessage>
+        )}
       </div>
     );
   },
