+++ packages/react/src/Select/stories/Select.stories.tsx
@@ -1,9 +1,24 @@
-import { Select } from '../';
+import { Select, SelectProps } from '../';
 
 const metadata = {
   title: 'Select',
   parameters: {
     layout: 'padded',
+  },
+  argTypes: {
+    description: {
+      defaultValue: 'Så vi kan hjelpe deg bedre',
+      control: 'text',
+    },
+    size: {
+      defaultValue: 'medium',
+      options: ['medium', 'small'],
+      control: { type: 'radio' },
+    },
+    required: {
+      defaultValue: true,
+      control: 'boolean',
+    },
   },
 };
 export default metadata;
@@ -16,16 +31,16 @@ const options = [
   { key: 'tromso', value: 'Tromsø' },
 ];
 
-export const Default = () => {
+export const Default = (props: SelectProps) => {
+  console.log(props);
   return (
-    <div className="flex flex-col gap-4">
-      <Select>
-        {options.map((option) => (
-          <option key={option.key} value={option.key}>
-            {option.value}
-          </option>
-        ))}
-      </Select>
-    </div>
+    <Select {...props} label="Velg område">
+      <option value="">Velg område</option>
+      {options.map((option) => (
+        <option key={option.key} value={option.key}>
+          {option.value}
+        </option>
+      ))}
+    </Select>
   );
 };
